requirements.txt
Django==4.1.7
django-rest-framework
psycopg2==2.9.6
